refactor: merge getFeedback and getLikeAndDislikeAmount into getDataFromStorage
