+++ pages/api/like.js
@@ -2,8 +2,7 @@ import CORS from 'cors'
 import globalAPICall from '../../utils/api/globalAPICall'
 import publishMessage from '../../utils/api/publishMessage'
 import { runMiddleware } from '../../utils/api/share'
-import { getLikeAndDislikeAmount } from '../../utils/api/getLikeAndDislikeAmount'
-import { likeFormName, likeFieldName } from '../../utils/api/config'
+import { getLikeAndDislikeAmount } from '../../utils/api/getDataFromStorage'
 
 const cors = CORS({
   methods: ['HEAD', 'PUT'],
@@ -24,8 +23,9 @@ async function handler(req, res) {
     }
   }
 
+  // get like and dislike amount
   async function GET() {
-    const result = await getLikeAndDislikeAmount(likeFormName, likeFieldName)
+    const result = await getLikeAndDislikeAmount()
 
     if (typeof result === 'string') {
       res.status(400).json({ message: result })
